Allow us to attach Notice of Election doc to an election
Once an election id is created, allow users to attach a
Notice of Elections doc to it. On submit, archive it.
It won't be there forever.

If a Notice of Election doc already exists,
link to the original URL and archived version

Notice of Election doc may exist for either
individual elections or an election group.