add storage app
We're going to need some helpers for dealing with File IO
in principle we might want to use them in
'elections', 'organisations', 'core' etc

Lets start putting them in a 'storage' app.

There is probably some other File IO code peppered about
in management commands that we could move here.
I probably won't refactor them on this branch, but
at least we've got a place for them to go now :)