improve router definitions/linking
This defines the routes in a better way,
removes the api/organisations/{pk} route
and makes the API more linkable.

If we want to link to an organisation record
from another endpoint, we may still need to
explicitly write:

organisation = OrganisationHyperlinkedIdentityField(
    view_name='organisation-detail', read_only=True
)

but this is definitely better.

Also adds some more useful fields to
the Organisation serializers

Also DRY up serializer definitions