add ElectedRoles model to /admin
In order to create elections relating to an org
we need an ElectedRole object for that org

It is also necessary to define a custom representation
of Organisation objects for use in /admin.

In the front-end we just want to show the org name,
but if we have a big drop-down of all the orgs
(with no date filter) we also need to show the dates
so we can disambiguate orgs with the same name.