data migration to ensure we have all official_identifiers
There are some wards in Cardiff where we updated
geography_curie with GSS code, but not official_identifier.

This migration just copies the GSS codes to the
official_identifier field so we don't lose that data.

We never used the pseudo-ids in YNR
so we're not losing anything we need.