change one line of code (which changes everything)
This is mostly explained in the lengthy comment,
but using public_objects as the default created
strange "hey where did my data go??" bugs.

For example, once we soft-deleted a record,
its visible it in /admin
(because I set it in get_queryset() )
but we could never un-delete it again because it was
invisible when selecting a related object.

Also dumpdata can only use the default manager and
we rely on that to sync data to "client" installs