Provide a way to get child elections with a specified manager
There isn't really a nice way to automatically apply the
filter of the parent manager through the child relationship.
The approach I've taken here removes the `.children`
property. (I've moved it to a _private property)
This forces us to make a choice about which manager
we are going to use when we access child records.
This seems like a reasonable tradeoff.
It is broadly consistent with what we're doing at the
top level of the hierarchy and doesn't introduce a lot
of additional overhead or unnecessary scope for error.