Update ballots when an Election is updated
When an 'parent' election is saved, we want those changes to be
reflected on the Election in YNR. So when we save an update, look for
any ballots and if they exist, update the modified timestamp on them.
This ensures that when the YNR importer runs looking for recent
changes it finds affected ballots and updates their parent Election.