Add a map to election templates
Only show it if there is a geography associated with an election object.
This means that election groups with an associated object (eg local
authorities /elections/local.aberdeen-city.2022-05-05/) will also show a
map.

To avoid cluttering the page, hide it in a details tag. Leaflet needs
the div with the target id to be added to the dom before it's
initialised. This is the reason for the eventListener on the detail.

Includes a not very informative popup that can be improved upon.