Take voting system from uk-election-ids
This change also removes "default" voting systems for top level
elections. Basically, we don't assign a voting system to `local.date`
because that sort of doesn't make sense.

If there is an organisation, we can make a guess at the voting system,
even if in some cases a nation might change that.