Don't pipe directly to S3
The S3 pipe is handy, but the pipe is opened _before_ the Postgres
connection (this is a UNIX thing). This means that if anything goes
wrong with the connection or the dump, the pipe ends up writing an empty
file to S3.

Later, this breaks any script that uses `pg_restore` on the empty file.