Tiny performance improvements
There are a number of things going on here.

Most of them make a tiny difference to the query performance, for
example `order_by()` on the queryset shaves off a few ms from each
request.

The main change is not counting rows when `coords` is passed in. This
saves performing the query twice (see existing comment in `list()`) and
was previously missing.

We only make a request to `coords` after showing an address picker on
WDIV, but we do that enough to make a difference here.

Tested with:

`ab -n 200 http://127.0.0.1:8000/api/elections/\?coords\=52.3831246,1.1102245\&future\=1\&current\=1\&identifier_type\=ballot`