Order API results by election group
There was a bug when paging through a lot of election results where the
order of the pages changed between requets. This meant that we coulnd't
build a full election tree when consuming results (e.g some pages ended
up duplicated, some ended up missing).

This change means that pages should be ordered in the same way when
paging.