install uv in the global env with pip instead of curl bashing
By default the curl installer installs into $HOME/.cargo/bin
this makes it a bit awakward to install as one user and then
call it as another.

You can work around this, but given uv has zero python
dependencies, installing it into the global env with pip won't
conflict with anything and saves some faffing about with paths
and permissions.

Installing it as root into the system env makes it available if
we log in with either the ubuntu or every_election user account